requirements.txt
optimum[onnxruntime]==1.2.3
mkl-include
mkl

seaborn
streamlit
pandas
matplotlib
plotly
pillow
scikit-learn